requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
common==0.1.2
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
Django==4.2.11
django-phonenumber-field==7.3.0
fonttools==4.49.0
html5lib==1.1
idna==3.6
lxml==5.1.0
oscrypto==1.3.0
phonenumbers==8.13.31
pillow==10.2.0
premailer==3.10.0
pycparser==2.21
pydyf==0.9.0
pyHanko==0.23.0
pyhanko-certvalidator==0.26.3
pypdf==4.1.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
xhtml2pdf==0.2.15
zopfli==0.2.3
